% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{calc_water_level_relative}
\alias{calc_water_level_relative}
\title{Calculate relative water levels}
\usage{
calc_water_level_relative(dat, r_means, prefix)
}
\arguments{
\item{dat}{Water level data.}

\item{r_means}{Moving-average reference table.}

\item{prefix}{File prefix.}
}
\value{
Relative water-level data.
}
\description{
Calculate relative water levels
}
\examples{
\dontrun{
calc_water_level_relative(dat, r_means, "data/data_karup")
}
}
