% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_main.R
\name{calcWaterLevelRelative}
\alias{calcWaterLevelRelative}
\title{Calc relative water levels}
\usage{
calcWaterLevelRelative(dat, rMeans, prefix)
}
\arguments{
\item{dat}{Water level records.}

\item{rMeans}{Average levels (tibble).}

\item{prefix}{Path prefix (e.g. data/data_skjern).}
}
\value{
The dataset.
}
\description{
Calc relative water levels
}
