% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_main.R
\name{fixStringErrors}
\alias{fixStringErrors}
\title{Estimate weight given dat}
\usage{
fixStringErrors(strings)
}
\arguments{
\item{strings}{Strings to fix}

\item{fn}{The csv file to save to.}

\item{dat}{The data set with column Period}
}
\value{
Fix UTF8 string error

Fixed strings
}
\description{
Estimate weight given dat
}
