% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_catch_table}
\alias{get_catch_table}
\title{Prepare catch records for table}
\usage{
get_catch_table(datCatch, datWeight)
}
\arguments{
\item{datCatch}{Catch table.}

\item{datWeight}{Weight estimates for missing data.}
}
\value{
The catch record table.
}
\description{
Prepare catch records for table
}
