% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_main.R
\name{getAllCatches}
\alias{getAllCatches}
\title{Read all catches from a river}
\usage{
getAllCatches(prefix)
}
\arguments{
\item{prefix}{File prefix such as 'data_karup_catch_seatrout'.}
}
\value{
The data.
}
\description{
Read all catches from a river
}
