% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_add_lines}
\alias{map_add_lines}
\title{Add line data to a Leaflet map}
\usage{
map_add_lines(map, group, data, color, useClub = TRUE)
}
\arguments{
\item{map}{A Leaflet map.}

\item{group}{Layer group name.}

\item{data}{Line data containing coordinates and popup fields.}

\item{color}{Line color.}

\item{useClub}{Whether club names should be included in layer names.}
}
\value{
A list containing the updated map and added group names.
}
\description{
Add line data to a Leaflet map
}
\examples{
\dontrun{
map_add_lines(leaflet(), "Fishing water", line_data, "#3C8AE6")
}
}
