% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_add_markers}
\alias{map_add_markers}
\title{Add marker data to a Leaflet map}
\usage{
map_add_markers(map, group, data)
}
\arguments{
\item{map}{A Leaflet map.}

\item{group}{Layer group name.}

\item{data}{Marker data containing coordinates and popup fields.}
}
\value{
A list containing the updated map and added group names.
}
\description{
Add marker data to a Leaflet map
}
\examples{
\dontrun{
map_add_markers(leaflet(), "Places", marker_data)
}
}
