% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_strip_kml}
\alias{map_strip_kml}
\title{Parse KML map data into marker/line tables}
\usage{
map_strip_kml(
  map_id,
  club = NA,
  group_name_markers = NULL,
  group_name_lines = NULL,
  start_ctr = 0
)
}
\arguments{
\item{map_id}{Google map ID.}

\item{club}{Optional club label.}

\item{group_name_markers}{Optional marker group name filter.}

\item{group_name_lines}{Optional line group name filter.}

\item{start_ctr}{Counter for line group IDs used up till now (to avoid duplicates when combining multiple maps).}
}
\value{
List with \code{datMarkers} and \code{datLines}.
}
\description{
Parse KML map data into marker/line tables
}
\examples{
\dontrun{
map_strip_kml("1XJoAUKY_-kbmhZgovPpLgi82Gn8")
}
}
