% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map_wrap_popup_content}
\alias{map_wrap_popup_content}
\title{Wrap map popup content}
\usage{
map_wrap_popup_content(x)
}
\arguments{
\item{x}{Character vector containing popup HTML.}
}
\value{
HTML suitable for a Leaflet popup.
}
\description{
Wrap map popup content
}
\examples{
map_wrap_popup_content("Fishing area")
}
