% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_reports.R
\name{mergeLists}
\alias{mergeLists}
\title{Merge two lists to one}
\usage{
mergeLists(a, b)
}
\arguments{
\item{a}{First list.}

\item{b}{Second list (priority).}
}
\description{
Merge two lists to one
}
