% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{monthly_stat_karup}
\alias{monthly_stat_karup}
\title{Calculate monthly catch statistics for Karup}
\usage{
monthly_stat_karup(dat_catch, year)
}
\arguments{
\item{dat_catch}{Catch records.}

\item{year}{Year under consideration.}
}
\value{
Monthly summary statistics.
}
\description{
Calculate monthly catch statistics for Karup
}
\examples{
\dontrun{
monthly_stat_karup(dat_catch, 2025)
}
}
