% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_reports.R
\name{monthlyStat}
\alias{monthlyStat}
\title{Calculate monthly catch statistics}
\usage{
monthlyStat(datCatch, year)
}
\arguments{
\item{datCatch}{Catch records.}

\item{year}{Year under consideration.}
}
\value{
Monthly catch statistics.
}
\description{
Calculate monthly catch statistics
}
