% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_catch}
\alias{read_catch}
\title{Read and prepare catch records for table}
\usage{
read_catch(prefix, dat_weight)
}
\arguments{
\item{prefix}{Prefix to csv files e.g. \code{"../../data/data_skjern_catch_salmon"}.}

\item{dat_weight}{Weight estimates.}
}
\value{
The catch records.
}
\description{
Read and prepare catch records for table
}
