% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_data}
\alias{read_data}
\title{Read data files from the \code{data} subfolder}
\usage{
read_data(
  file_prefix,
  year = NULL,
  path = "https://raw.githubusercontent.com/relund/riverdata/master/data/"
)
}
\arguments{
\item{file_prefix}{File prefix, e.g. "data_skjern_catch_salmon_".}

\item{year}{Year(s) to consider.}

\item{path}{Path to data folder.}
}
\value{
The data (tibble).
}
\description{
Read data files from the \code{data} subfolder
}
\examples{
\dontrun{
read_data(file_prefix = "data_skjern_catch_salmon_", year = 2023:2024)
}
}
