% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_data_files}
\alias{read_data_files}
\title{Read and combine data files by filename pattern}
\usage{
read_data_files(pattern, path = NULL)
}
\arguments{
\item{pattern}{Regex pattern used to match files in \verb{data/}.}

\item{path}{Optional base data path. If \code{NULL}, a sensible relative path is
auto-detected from the current working directory.}
}
\value{
Combined data table.
}
\description{
Read and combine data files by filename pattern
}
\examples{
\dontrun{
read_data_files("data_karup_waterlevel_[0-9]{4}")
}
}
