% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_w_temp}
\alias{read_w_temp}
\title{Read yearly water temperature files}
\usage{
read_w_temp(prefix, years)
}
\arguments{
\item{prefix}{File prefix.}

\item{years}{Integer vector of years.}
}
\value{
Combined water temperature data.
}
\description{
Read yearly water temperature files
}
\examples{
\dontrun{
read_w_temp("data/data_karup", 2020:2025)
}
}
