% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_reports.R
\name{readCatch}
\alias{readCatch}
\title{Read and prepare catch records for table}
\usage{
readCatch(prefix, datWeight)
}
\arguments{
\item{prefix}{Prefix to csv files e.g. "../../data/data_skjern_catch_salmon".}

\item{datWeight}{Weight estimates.}
}
\value{
The catch records
}
\description{
Read and prepare catch records for table
}
