% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_reports.R
\name{readCatchDTU}
\alias{readCatchDTU}
\title{Read catch records from a DTU json file}
\usage{
readCatchDTU(url)
}
\arguments{
\item{url}{Url for json (without year) ending with a slash.}
}
\value{
The data (tibble).
}
\description{
Read catch records from a DTU json file
}
