% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_main.R
\name{readDataFiles}
\alias{readDataFiles}
\title{Read data files from the \code{data} subfolder}
\usage{
readDataFiles(pattern)
}
\arguments{
\item{pattern}{File to search for (can use regexp) (e.g. 'data_karup_catch_seatrout_[0-9]{4}').}
}
\value{
The data (tibble).
}
\description{
Read data files from the \code{data} subfolder
}
