% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_main.R
\name{readWTemp}
\alias{readWTemp}
\title{Update and save water level data for current year}
\usage{
readWTemp(prefix, years)
}
\arguments{
\item{prefix}{Path prefix (e.g. data/data_skjern).}

\item{years}{Years to load.}

\item{days}{Number of days back to read.}

\item{stations}{Stations under consideration.}
}
\value{
Range of dates.
Update and save water level data for current year

The dataset.
#' Read water level files
#'
#' @param prefix Path prefix (e.g. data/data_skjern).
#' @param years Years to load.
#'
#' @return The data set.
readWLevels <- function(prefix, years) {
dat <- NULL
for (y in years) {
fn <- paste0(prefix, "\emph{waterlevel}", y, ".csv")
dat <-
bind_rows(dat,
read_csv(fn, col_types = cols(
Date = col_datetime(format = ""),
Place = col_character(),
Value = col_double()
)))
}
return(dat)
}
Read water temperature files

The data set.
}
\description{
Update and save water level data for current year
}
\note{
Stations can be found at http://www.hydrometri.dk/hyd/. Get obs for the last 100 days

Stations can be found at https://vandportalen.dk/.
#' @return Range of dates.
updateWaterLevel_new <- function(stations, prefix, days = 100) {
iso <- format(now(), format = "\%Y-\%m-\%dT\%T.111Z", tz = "GMT")
dat <- NULL
for (i in 1:nrow(stations)) {
id <- stations$id\link{i}
place <- stations$place\link{i}
# get obs for last 100 days
tmp <- fromJSON(paste0("https://vandportalen.dk/api/hyd/getplotdata?tsid=", id, "&enddate=", iso, "&days=", days, "&pw=100000000&inclraw=true"))
offset <- as.numeric(tmp$tsh$Offset)
tmp <- as_tibble(tmp$PlotRecs\link{,1:2}) \%>\% mutate(V = sapply(tmp$PlotRecs\link{,2}, function(x) {x\link{1}}))
tmp$V <- tmp$V - rep(offset, length(tmp$V))
colnames(tmp) <- c("Date", place)
if (is.null(dat)) {
dat <- tmp
} else {
dat <- full_join(dat, tmp, by = "Date")
}
}
dat$Date <- ymd_hms(dat$Date, tz = "UTC") # \%>\% with_tz("CET") # from UTC to CET
dat <- dat \%>\%
pivot_longer(cols = 2:ncol(dat), names_to = 'Place', values_to = 'Level') \%>\%
filter(!is.na(Level))save to files

for (y in unique(year(dat$Date))) {
fn <- paste0(prefix, "\emph{waterlevel}", y, ".csv")
# if (file.exists(fn)) {
#   datOld <- read_csv(fn, col_types = "Tcd")
# } else
datOld <- NULL
message("Waterlevel: Update dataset for year ", y)
dat1 <- dat \%>\% dplyr::filter(year(Date) == y) \%>\%
arrange_all(desc) \%>\%
distinct(Date, .keep_all = T)
# combine with old data
dat1 <- bind_rows(datOld, dat1) \%>\%
arrange_all(desc) \%>\%
distinct(Date, Place, .keep_all = T) \%>\%
select(Date, Place, Level)
# remove outliers
dat1 <- as_tsibble(dat1, key = Place, index = Date) \%>\%
group_by_key() \%>\%
mutate(Level = tsclean(Level, replace.missing = FALSE)) \%>\%
as_tibble()
# save
message("Write data to ", fn)
write_csv(dat1, fn)
}
return(range(dat$Date))
}
Get and save all water level data for 2012 until now

Stations can be found at https://vandportalen.dk/ (look at the url).
}
