% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snip.R
\name{snip_leaflet}
\alias{snip_leaflet}
\title{Build an interactive river map}
\usage{
snip_leaflet(prefix, show_groups = "Stednavne")
}
\arguments{
\item{prefix}{File prefix for the map marker and line CSV files.}

\item{show_groups}{Layer groups shown when the map opens.}
}
\value{
A Leaflet map.
}
\description{
Build an interactive river map
}
\examples{
\dontrun{
snip_leaflet("data/data_karup_", show_groups = "Stednavne")
}
}
