% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snip.R
\name{snip_table}
\alias{snip_table}
\title{Catch summary table}
\usage{
snip_table(datCatch, .row_unit = c("Year"))
}
\arguments{
\item{datCatch}{Catch data.}

\item{.row_unit}{Row unit for the summary table.}
}
\value{
A formatted HTML table.
}
\description{
Catch summary table
}
\examples{
dat <- read_data("data_karup_catch_seatrout_", year = 2023)
snip_table(dat)
}
