% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_main.R
\name{splitDataFileByYear}
\alias{splitDataFileByYear}
\title{Split a data file into year files}
\usage{
splitDataFileByYear(path, prefix)
}
\arguments{
\item{path}{Path to file e.g. "data/data_karup_catch_seatrout_2020-2022.csv".}

\item{prefix}{Prefix path for saved files e.g. "data/data_karup_catch_seatrout".}
}
\description{
Split a data file into year files
}
