% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_lock_skjern}
\alias{write_lock_skjern}
\title{Write Skjern lock flow data}
\usage{
write_lock_skjern(prefix)
}
\arguments{
\item{prefix}{File prefix.}
}
\value{
Lock flow table.
}
\description{
Write Skjern lock flow data
}
\examples{
\dontrun{
write_lock_skjern("data/data_skjern")
}
}
