% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_lock_web}
\alias{write_lock_web}
\title{Write lock web dataset}
\usage{
write_lock_web(dat, prefix)
}
\arguments{
\item{dat}{Flow data.}

\item{prefix}{File prefix.}
}
\value{
Invisibly returns transformed data.
}
\description{
Write lock web dataset
}
\examples{
\dontrun{
write_lock_web(dat, "data/data_skjern")
}
}
