% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_water_levels_web}
\alias{write_water_levels_web}
\title{Write water level web dataset}
\usage{
write_water_levels_web(dat, prefix)
}
\arguments{
\item{dat}{Water level data.}

\item{prefix}{File prefix.}
}
\value{
Web dataset.
}
\description{
Write water level web dataset
}
\examples{
\dontrun{
write_water_levels_web(dat, "data/data_skjern")
}
}
