% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_water_mov_avg}
\alias{write_water_mov_avg}
\title{Write water level moving averages}
\usage{
write_water_mov_avg(dat, prefix)
}
\arguments{
\item{dat}{Water level data.}

\item{prefix}{File prefix.}
}
\value{
Moving-average table.
}
\description{
Write water level moving averages
}
\examples{
\dontrun{
write_water_mov_avg(dat, "data/data_karup")
}
}
