% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_water_temp_web}
\alias{write_water_temp_web}
\title{Write water temperature web dataset}
\usage{
write_water_temp_web(dat, r_means, prefix)
}
\arguments{
\item{dat}{Water temperature data.}

\item{r_means}{Moving-average reference table.}

\item{prefix}{File prefix.}
}
\value{
Web dataset.
}
\description{
Write water temperature web dataset
}
\examples{
\dontrun{
write_water_temp_web(dat, r_means, "data/data_skjern")
}
}
