% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_weight_estimates}
\alias{write_weight_estimates}
\title{Write weight estimates}
\usage{
write_weight_estimates(prefix, seatrout = TRUE)
}
\arguments{
\item{prefix}{File prefix.}

\item{seatrout}{Whether to estimate seatrout (\code{TRUE}) or salmon (\code{FALSE}).}
}
\value{
Weight estimate table.
}
\description{
Write weight estimates
}
\examples{
\dontrun{
write_weight_estimates("data/data_skjern", seatrout = TRUE)
}
}
