% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_main.R
\name{writeCatch}
\alias{writeCatch}
\title{Save catch records from a given year to a csv file}
\usage{
writeCatch(url, prefix, yr, species = "Havørred", club = FALSE)
}
\arguments{
\item{url}{Url for json (without year) ending with a slash.}

\item{prefix}{Prefix for csv files such as 'data/data_karup'.}

\item{yr}{Year to get data for.}

\item{species}{Species. Either "Havørred" or "Laks".}

\item{club}{False if consider association. True if consider club.}
}
\value{
The data (tibble).
}
\description{
Save catch records from a given year to a csv file
}
