% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_main.R
\name{writeCatchKarup}
\alias{writeCatchKarup}
\title{Save catch records from 2020 to present to csv file}
\usage{
writeCatchKarup()
}
\value{
The data.
}
\description{
Save catch records from 2020 to present to csv file
}
