% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_main.R
\name{writeLockSkjern}
\alias{writeLockSkjern}
\title{Update and save lock data for Hvide Sande}
\usage{
writeLockSkjern(prefix)
}
\arguments{
\item{prefix}{Path prefix (e.g. data/data_skjern).}
}
\value{
The lock data set.
}
\description{
Update and save lock data for Hvide Sande
}
