% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_main.R
\name{writeLockWeb}
\alias{writeLockWeb}
\title{Calc flow dataset for web}
\usage{
writeLockWeb(dat, prefix)
}
\arguments{
\item{dat}{Flow data.}

\item{prefix}{Path prefix (e.g. data/data_skjern).}
}
\value{
The dataset with column \code{Open} equal 0 (closed), 1 (partly open) and 2 (open).
}
\description{
Calc flow dataset for web
}
