% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_main.R
\name{writeTimeSeriesData}
\alias{writeTimeSeriesData}
\title{Update and save time series data from vandportalen.dk}
\usage{
writeTimeSeriesData(stations = NULL, prefix, prefix1, days)
}
\arguments{
\item{stations}{Stations under consideration.}

\item{prefix}{Path prefix (e.g. data/data_skjern).}

\item{prefix1}{Prefix to append to \code{prefix} (e.g. waterlevel).}

\item{days}{Number of days back to get data for.}
}
\value{
Data set (tibble).
}
\description{
Update and save time series data from vandportalen.dk
}
\note{
Stations can be found at https://vandportalen.dk/ (look at the url).
}
