% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_main.R
\name{writeWaterMovAvg}
\alias{writeWaterMovAvg}
\title{Calc and save moving average for water level}
\usage{
writeWaterMovAvg(dat, prefix)
}
\arguments{
\item{dat}{Water level records.}

\item{prefix}{Path prefix (e.g. data/data_skjern).}
}
\value{
The data set.
}
\description{
Calc and save moving average for water level
}
