% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_main.R
\name{writeWaterTempMovAvg}
\alias{writeWaterTempMovAvg}
\title{Calc moving average for water temperature}
\usage{
writeWaterTempMovAvg(dat, prefix)
}
\arguments{
\item{dat}{Water temperature records.}

\item{prefix}{Path prefix (e.g. data/data_skjern).}
}
\value{
The data set.
}
\description{
Calc moving average for water temperature
}
