% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_main.R
\name{writeWaterTempWeb}
\alias{writeWaterTempWeb}
\title{Calculate dataset for web}
\usage{
writeWaterTempWeb(dat, rMeans, prefix)
}
\arguments{
\item{dat}{Water level data set.}

\item{prefix}{Path prefix (e.g. data/data_skjern).}
}
\value{
The data set
}
\description{
Calculate dataset for web
}
