% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_main.R
\name{writeWeightEstimates}
\alias{writeWeightEstimates}
\title{Estimate weight ranges.}
\usage{
writeWeightEstimates(prefix, seatrout = TRUE)
}
\arguments{
\item{prefix}{Prefix for csv files such as 'data/data_karup'.}

\item{seatrout}{If true get data for seatrout otherwise salmon.}
}
\value{
The estimated results (tibble).
}
\description{
Estimate weight ranges.
}
