% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{yearly_stat}
\alias{yearly_stat}
\title{Calculate yearly catch statistics}
\usage{
yearly_stat(dat_catch)
}
\arguments{
\item{dat_catch}{Catch records.}
}
\value{
Yearly summary statistics.
}
\description{
Calculate yearly catch statistics
}
\examples{
\dontrun{
yearly_stat(dat_catch)
}
}
