% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{yearly_stat_karup}
\alias{yearly_stat_karup}
\title{Calculate yearly catch statistics for Karup}
\usage{
yearly_stat_karup(dat_catch)
}
\arguments{
\item{dat_catch}{Catch records.}
}
\value{
Yearly summary statistics.
}
\description{
Calculate yearly catch statistics for Karup
}
\examples{
\dontrun{
yearly_stat_karup(dat_catch)
}
}
