% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_reports.R
\name{yearlyStatKarup}
\alias{yearlyStatKarup}
\title{Calculate yearly catch statistics}
\usage{
yearlyStatKarup(datCatch)
}
\arguments{
\item{datCatch}{Catch records.}
}
\value{
Yearly catch statistics.
}
\description{
Calculate yearly catch statistics
}
