% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverdata-package.R
\docType{package}
\name{riverdata-package}
\alias{riverdata}
\alias{riverdata-package}
\title{riverdata: R package used by riverdata repo}
\description{
Package for riverdata repo at Github.
}
\author{
\strong{Maintainer}: Lars Relund Nielsen \email{lars@relund.dk} (\href{https://orcid.org/0000-0002-4802-3071}{ORCID})

}
\keyword{internal}
