% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snip.R
\name{snip_plot_catch}
\alias{snip_plot_catch}
\title{Plot of catches given current date}
\usage{
snip_plot_catch(
  datCatch,
  .look_back = 30,
  .unit = c("cur_year_days", "cur_year_weeks", "cur_year_months", "months", "years"),
  .group = c("Place", "Method"),
  .plotly = TRUE
)
}
\arguments{
\item{datCatch}{Catch data.}

\item{.look_back}{Number of time units to plot.}

\item{.unit}{Time unit for the look-back window. One of \code{"cur_year_days"}, \code{"cur_year_weeks"}, \code{"cur_year_months"}, \code{"months"} or \code{"years"}.}

\item{.group}{Column name used for grouping and fill in the bar plot. One of \code{"Place"} or \code{"Method"}.}

\item{.plotly}{Use plotly otherwise just ggplot.}
}
\value{
The plot object.
}
\description{
Plot of catches given current date
}
\examples{
dat <- read_data("data_karup_catch_seatrout_", year = 2023)
snip_plot_catch(dat)
}
