% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snip.R
\name{snip_plot_catch}
\alias{snip_plot_catch}
\title{Plot of catches given current date}
\usage{
snip_plot_catch(datCatch, .days = 30, .plotly = TRUE)
}
\arguments{
\item{datCatch}{Catch data.}

\item{.days}{Number of days to plot.}

\item{.plotly}{Use plotly otherwise just ggplot.}
}
\value{
The plot object.
}
\description{
Plot of catches given current date
}
\examples{
dat <- read_data("data_karup_catch_seatrout_", year = 2023)
snip_plot_catch(dat)
}
