% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_catch}
\alias{write_catch}
\title{Save fangstjournalen catch records from a given year to a csv file}
\usage{
write_catch(url, prefix, yr, species = "seatrout", club = FALSE, write = TRUE)
}
\arguments{
\item{url}{Url for json (without year) ending with a slash.}

\item{prefix}{Prefix path for the csv file.}

\item{yr}{Year to get data for (single number).}

\item{species}{Species. Either "seatrout" or "salmon".}

\item{club}{False if consider association. True if consider club.}

\item{write}{Whether to write the result to a CSV file.}
}
\value{
The data (tibble).
}
\description{
Data is written to \code{str_c(prefix, "_catch_", res, "_", yr, ".csv")}.
}
\examples{
\dontrun{
url <- str_c("https://fangstjournalen.dtu.dk/fangst.nsf/xsp/app/v3/",
             "catches/assoc/A97F957DD48AEDD4C1258814003E71FE/1/")
prefix <- "tmp/data_skjern"
yr <- 2023
write_catch(url, prefix, yr, species = "Laks")
}

}
