% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{writeCatch}
\alias{writeCatch}
\title{Save fangstjournalen catch records from a given year to a csv file}
\usage{
writeCatch(url, prefix, yr, species = "Havørred", club = FALSE)
}
\arguments{
\item{url}{Url for json (without year) ending with a slash.}

\item{prefix}{Prefix path for the csv file.}

\item{yr}{Year to get data for (single number).}

\item{species}{Species. Either "Havørred" or "Laks".}

\item{club}{False if consider association. True if consider club.}
}
\value{
The data (tibble).
}
\description{
Data is written to \code{str_c(prefix, "_catch_", res, "_", yr, ".csv")}.
}
\examples{
url <- "https://fangstjournalen.dtu.dk/fangst.nsf/xsp/app/v3/catches/assoc/A97F957DD48AEDD4C1258814003E71FE/1/"
prefix <- "data/data_skjern"
yr <- year(now())
datCatchSalmon <- writeCatch(url, prefix, yr, species = "Laks")

}
